General update on A3; clarification of output and execution.
A2: Added a test to verify that the solution meets the Fs>1.
As I've seen an reasonable amount of solutions that use integer based APIs, this
will cause problems when you try to time/sleep fractions of seconds. However, as the
solution needs to hande Fs>1 (i.e. Ts<1 s) you need to be able to observer time at such
resolution. You also need to 'sleep' with similar capabilities.